Refactored code. Added new translations
Added new keyboard referral_links_bk, renamed back_to_main_kb. Refactored code to improve interaction
