+++ bot/config.py
@@ -21,6 +21,7 @@ dp = Dispatcher(storage=storage)
 SUPPORTED_LANGUAGES = ['ru', 'sk', 'en', 'uk', 'es', 'fr', 'tr', 'ar']
 
 REFERRAL_LINKS = {
+    '🐕 Lost Dogs: The Way': 'https://t.me/lost_dogs_bot/lodoapp?startapp=ref-u_452792586__s_629960',
     'Cats 🐈‍⬛': 'https://t.me/catsgang_bot/join?startapp=Lv39AXNhcBvwK6ZWcoGpD',
     '😺 Catizen': 'https://t.me/catizenbot/gameapp?startapp=r_2779_5703457',
     '🏠 City Holder': 'https://t.me/blum/app?startapp=ref_8U6nYohQav',
