+++ bot/handlers/handlers.py
@@ -13,12 +13,13 @@ from database.database import (get_session, get_or_create_user, update_user_lang
                                get_subscribed_users, get_user_role_and_ban_info, update_safety_keys_generated,
                                delete_safety_keys, get_safety_keys, check_user_limits, check_user_safety_limits)
 
-from keyboards.back_to_main import get_back_to_main_menu_button
+from keyboards.back_to_main_kb import get_back_to_main_menu_button
+from keyboards.referral_links_kb import referral_links_keyboard
 from keyboards.inline import (get_action_buttons, get_settings_menu, create_language_keyboard,
                               back_to_main_menu_key, get_admin_panel_keyboard, get_main_in_admin,
                               get_detail_info_in_admin,
                               notification_menu, confirmation_button_notification,
-                              instruction_prem_button, referral_links_keyboard)
+                              instruction_prem_button)
 from utils.helpers import load_translations, get_translation, get_remaining_time
 from states.form import Form, FormSendToUser
 
@@ -121,8 +122,8 @@ async def send_keys_menu(message: types.Message, state: FSMContext):
         )
 
 
-@dp.callback_query(F.data == "my_referral_links")
-async def get_my_referral_links(callback_query: types.CallbackQuery):
+@dp.callback_query(F.data == "referral_links")
+async def referral_links_handler(callback_query: types.CallbackQuery):
     async with await get_session() as session:
         user_id = callback_query.from_user.id if callback_query.from_user.id != BOT_ID else callback_query.chat.id
         image_dir = os.path.join(os.path.dirname(__file__), "..", "images", "premium")
@@ -135,27 +136,28 @@ async def get_my_referral_links(callback_query: types.CallbackQuery):
                 photo = FSInputFile(image_path)
                 caption_ref_link = await get_translation(user_id, "referral_links_description")
                 new_media = types.InputMediaPhoto(media=photo, caption=caption_ref_link)
+                chat_id=callback_query.message.chat.id
+                message_id = callback_query.message.message_id
+                keyboard = await referral_links_keyboard(user_id)
 
                 if callback_query.message.photo:
                     await bot.edit_message_media(
-                        chat_id=callback_query.message.chat.id,
-                        message_id=callback_query.message.message_id,
+                        chat_id=chat_id,
+                        message_id=message_id,
                         media=new_media,
-                        reply_markup=referral_links_keyboard()
+                        reply_markup=keyboard
                     )
                 else:
                     await bot.delete_message(
-                        chat_id=callback_query.message.chat.id,
-                        message_id=callback_query.message.message_id
+                        chat_id=chat_id,
+                        message_id=message_id
                     )
                     await bot.send_photo(
-                        chat_id=callback_query.message.chat.id,
+                        chat_id=chat_id,
                         photo=photo,
                         caption=caption_ref_link,
-                        reply_markup=referral_links_keyboard()
+                        reply_markup=keyboard
                     )
-
-
 
 
 async def execute_change_language_logic(message: types.Message, user_id: int, state: FSMContext):
@@ -1020,21 +1022,21 @@ async def show_info_message(callback_query: types.CallbackQuery):
         chat_id = callback_query.message.chat.id
         message_id = callback_query.message.message_id
         info_caption = await get_translation(user_id, "info_message")
-        main_menu_back = await get_back_to_main_menu_button(user_id)
+        keyboard = await get_back_to_main_menu_button(user_id)
 
         if callback_query.message.photo:
             await bot.edit_message_caption(
                 chat_id=chat_id,
                 message_id=message_id,
                 caption=info_caption,
-                reply_markup=main_menu_back
+                reply_markup=keyboard
             )
         else:
             await bot.edit_message_text(
                 chat_id=chat_id,
                 message_id=message_id,
                 text=info_caption,
-                reply_markup=main_menu_back
+                reply_markup=keyboard
             )
